+++ arcore_plugin/example/lib/main.dart
@@ -1,7 +1,24 @@
+import 'dart:io';
+import 'package:flutter/services.dart' show rootBundle;
 import 'package:arcore_plugin/arcore_plugin.dart';
 import 'package:flutter/material.dart';
 
-void main() => runApp(MaterialApp(home: TextViewExample()));
+void main() async {
+  final Directory systemTempDir = Directory.systemTemp;
+  final File tempFile = File('${systemTempDir.path}/image_database.imgdb');
+
+  // create tempfile
+  await tempFile.create();
+
+  rootBundle.load("assets/image_database.imgdb").then((data) {
+    tempFile.writeAsBytesSync(
+        data.buffer.asUint8List(data.offsetInBytes, data.lengthInBytes));
+
+    runApp(MaterialApp(home: TextViewExample()));
+  }).catchError((error) {
+    throw Exception(error);
+  });
+}
 
 class TextViewExample extends StatefulWidget {
   @override
@@ -9,6 +26,8 @@ class TextViewExample extends StatefulWidget {
 }
 
 class _TextViewExampleState extends State<TextViewExample> {
+  String recongizedImage;
+  ArCoreViewController arCoreViewController;
 
   @override
   void initState() {
@@ -30,11 +49,20 @@ class _TextViewExampleState extends State<TextViewExample> {
                 width: screenSize.width,
                 height: screenSize.height,
                 child: ArCoreView(
+                  onImageRecognized: _onImageRecognized,
                   onArCoreViewCreated: _onTextViewCreated,
                 ))));
   }
 
   void _onTextViewCreated(ArCoreViewController controller) {
+    arCoreViewController = controller;
     controller.getArCoreView();
+  }
+
+  void _onImageRecognized(String recImgName) {
+    print("image recongized: $recImgName");
+
+    // you can pause the image recognition via arCoreViewController.pauseImageRecognition();
+    // resume it via arCoreViewController.resumeImageRecognition();
   }
 }