+++ arcore_plugin/CHANGELOG.md
@@ -1,3 +1,14 @@
 ## 0.0.1
 
 * First Release 
+
+## 0.0.1+1
+
+* Added more information in README file 
+
+## 0.0.1+2
+
+* Added better control over Image Recognition. User can pause and resume image recognition via ArCoreViewController
+* Added example on how to save the imgdb file in the App --> see main function in example 
+* Added pretrained imgdb file and tester images as asset for example 
+* updated README.md